+++ renamifier.h
@@ -20,10 +20,10 @@
 #ifndef RENAMIFIER_H
 #define RENAMIFIER_H
 
-#define VERSION "0.1.2"
+#define VERSION "0.1.3"
 
 #define COPYRIGHT_HTML \
-"<p>Copyright &copy; 2021, 2022 Benjamin Johnson</p>"
+"<p>Copyright &copy; 2021, 2022, 2023 Benjamin Johnson</p>"
 
 #define LICENSE_TEXT_HTML \
 "<p>This program is free software; you can redistribute it and/or modify " \