+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+* Remove obsolete dependencies when upgrading the Windows version.
+
 ## [0.1.2] - 2022-10-15
 ### Changed
 * Prompt to exit the application after rename-and-moving the last file.