+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+* Open the rename and move dialog in the file's original location the first time it is used.
 ### Fixed
 * Always preserve the original file extension when rename-and-moving.
 * Remove obsolete dependencies when upgrading the Windows version.